refactor: consolidate InvalidOperationHashError into @taquito/core and improve across packages
re #1994
