+++ packages/taquito/src/wallet/operation.ts
@@ -16,9 +16,10 @@ import {
 } from 'rxjs/operators';
 import { Context } from '../context';
 import { Receipt, receiptFromOperation } from './receipt';
-import { validateOperation, ValidationResult, InvalidOperationHashError } from '@taquito/utils';
+import { validateOperation, ValidationResult } from '@taquito/utils';
 import { BlockIdentifier } from '../read-provider/interface';
 import { InvalidConfirmationCountError, ConfirmationUndefinedError } from '../error';
+import { InvalidOperationHashError } from '@taquito/core';
 
 export type OperationStatus = 'pending' | 'unknown' | OperationResultStatusEnum;
 
@@ -87,6 +88,7 @@ export class WalletOperation {
    * @param opHash Operation hash
    * @param raw Raw operation that was injected
    * @param context Taquito context allowing access to rpc and signer
+   * @throws {InvalidOperationHashError}
    */
   constructor(
     public readonly opHash: string,
